Revert "Create marvel service"
