--- Marvel.xcodeproj/xcuserdata/spessotovitor.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "0C020244-F0EC-4B9F-A6DE-4567A4AE788B"
-   type = "1"
-   version = "2.0">
-</Bucket>
