+++ Marvel/Features/Heros/Models/UrlModel.swift
@@ -1,0 +1,14 @@
+//
+//  UrlModel.swift
+//  Marvel
+//
+//  Created by Vitor Spessoto on 10/02/22.
+//
+
+import Foundation
+
+struct UrlModel: Codable {
+
+    let type: String
+    let url: String
+}
