--- Marvel/Features/HerosList/Views/Cell/HerosCell.swift
@@ -1,63 +1,0 @@
-//
-//  HerosCell.swift
-//  Marvel
-//
-//  Created by Vitor Spessoto on 12/02/22.
-//
-
-import UIKit
-
-class HerosCell: UITableViewCell {
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-        // Configure the view for the selected state
-    }
-    
-    lazy var contactImage: UIImageView = {
-        let imgView = UIImageView()
-        imgView.translatesAutoresizingMaskIntoConstraints = false
-        imgView.contentMode = .scaleAspectFit
-        imgView.clipsToBounds = true
-        return imgView
-    }()
-    
-    lazy var fullnameLabel: UILabel = {
-        let label = UILabel()
-        label.font = UIFont.systemFont(ofSize: 20, weight: .semibold)
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        
-        configureViews()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        
-        configureViews()
-    }
-    
-    func configureViews() {
-        contentView.addSubview(contactImage)
-        contentView.addSubview(fullnameLabel)
-        
-        contactImage.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 15).isActive = true
-        contactImage.centerYAnchor.constraint(equalTo: contentView.centerYAnchor).isActive = true
-        contactImage.heightAnchor.constraint(equalToConstant: 100).isActive = true
-        contactImage.widthAnchor.constraint(equalToConstant: 100).isActive = true
-        
-        fullnameLabel.leadingAnchor.constraint(equalTo: contactImage.trailingAnchor, constant: 16).isActive = true
-        fullnameLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -15).isActive = true
-        fullnameLabel.topAnchor.constraint(equalTo: contentView.topAnchor).isActive = true
-        fullnameLabel.bottomAnchor.constraint(equalTo: contentView.bottomAnchor).isActive = true
-    }
-}
